Fix emit signal from object in other thread
The issue was that the object, which emitted the signal,
did was not created in the main thread. So when
the signal was emitted, the queued connection would try
to call the slot in the thread where the object was created in.
But only the main thread has a running event loop.

There are two options: Move the object to another thread.
Or connect to a slot that will definitly live in the main thread.
The second option is way easier.